auth-fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
certifi==2025.10.5
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2==2.9.11
pydantic==2.12.1
pydantic_core==2.41.3
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
tomli==2.3.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
